// client/src/utils/types.ts
export interface User {
  id: string;
  email: string;
  name: string;
  imageUrl: string | null;
}
export type LoggedInUser = {
  isAuthenticated: boolean;
  user: User | null;
};
export type AuthContextType = {
  loggedInUser: LoggedInUser;
  setLoggedInUser: React.Dispatch<SetStateAction<LoggedInUser>>;
  showLoading: boolean;
};
export interface SocketContextType {
  socket: Socket | null;
  // setSocket: React.Dispatch<SetStateAction<Socket | null>> | null;
}

export interface ConversationContextType {
  allMessages: Message[];
  messagesEndRef: React.RefObject<HTMLDivElement>;
  currentConversation: Conversation | null;
  setCurrentConversation: React.Dispatch<
    SetStateAction<Conversation | null>
  > | null;
  numberOfOnlineUsersInCurrentConversation: number;
  setNumberOfOnlineUsersInCurrentConversation: React.Dispatch<
    SetStateAction<number>
  >;
  handleGoToHome: () => void;
  newMessagesInConversations: Message[];
  setNewMessageInConversations: React.Dispatch<SetStateAction<Message[]>>;
  groupTitle: string;
  setGroupTitle: React.Dispatch<React.SetStateAction<string>>;
  conversations: Conversation[];
  allUsers: User[];
  addChatAnchorEl: HTMLElement | null;
  setAddChatAnchorEl: React.Dispatch<React.SetStateAction<HTMLElement | null>>;
  openCreateConversationModal: {
    isOpen: boolean;
    type: ConversationType;
  };
  setOpenCreateConversationModal: React.Dispatch<
    React.SetStateAction<{
      isOpen: boolean;
      type: ConversationType;
    }>
  >;
  selectedUserForConversation: User[];
  setSelectedUserForConversation: React.Dispatch<React.SetStateAction<User[]>>;
  handleCreateConversation: () => Promise<void>;
  handleSearchUserChange: (
    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
  ) => void;
  handleDeleteConversation: () => Promise<void>;
  chatMenuAnchorEl: HTMLElement | null;
  setChatMenuAnchorEl: React.Dispatch<React.SetStateAction<HTMLElement | null>>;
  searchConversationValue: string;
  setSearchConversationValue: React.Dispatch<React.SetStateAction<string>>;

  handleResetNewMessagesInConversation: (cId: string) => void;
  searchUserValue: string;
}

export interface Message {
  id: string;
  body: string;
  conversationId: string;
  senderId: string;
  senderName: string;
  senderImageUrl: string;
  senderEmail: string;
  createdAt: string;
  fileName?: string | null;
  fileUrl?: string | null;
}

export interface Sender {
  id: string;
  joinedAt: string;
  userId: string;
  conversationId: string;
  title: string;
  role: string;
  imageUrl: string;
}
// Generated by https://quicktype.io

export interface Conversation {
  id: string;
  isGroup: boolean;
  groupTitle: null;
  type: string;
  createdAt: string;
  updatedAt: string;
  members: Member[];
}
export interface Member {
  id: string;
  joinedAt: string;
  userId: string;
  conversationId: string;
  title: string;
  imageUrl: string;
  role: string;
}
// ...
export type ConversationType = "DIRECT_MESSAGE" | "GROUP";